adding correct toolpit text and cleaning
